Extract HTML tool card renderer
