+++ Sources/QuillCodeApp/WorkspaceHTMLRenderer.swift
@@ -290,7 +290,7 @@ public enum WorkspaceHTMLRenderer {
             """
         case .toolCard:
             guard let card = item.toolCard else { return "" }
-            return renderToolCard(card, timelineItemID: item.id)
+            return WorkspaceHTMLToolCardRenderer.render(card, timelineItemID: item.id)
         }
     }
 
@@ -409,179 +409,6 @@ public enum WorkspaceHTMLRenderer {
         """
     }
 
-    private static func renderToolCard(_ card: ToolCardState, timelineItemID: String? = nil) -> String {
-        let timelineAttribute = timelineItemID.map { #" data-timeline-id="\#(escape($0))""# } ?? ""
-        let executionContextAttribute = card.executionContext
-            .map { #" data-execution-context="\#(escape($0.kind.rawValue))""# } ?? ""
-        let accessibilityContext = card.executionContext
-            .map { ", \($0.label) \($0.detail)" } ?? ""
-        let copyID = timelineItemID ?? card.id
-        return """
-        <article class="tool-card \(card.status.rawValue)" data-testid="tool-card" data-status="\(card.status.rawValue)" data-density="\(card.density.rawValue)" aria-label="\(escape(card.title)), \(escape(card.status.rawValue)), \(escape(card.densityAccessibilityLabel))\(escape(accessibilityContext))"\(timelineAttribute)\(executionContextAttribute)>
-          <header>
-            <span class="tool-card-title-row">
-              <strong data-testid="tool-card-title">\(escape(card.title))</strong>
-              \(renderExecutionContextChip(card.executionContext, testID: "tool-card-execution-context"))
-            </span>
-            <span data-testid="tool-card-status">\(escape(card.status.rawValue))</span>
-          </header>
-          <p data-testid="tool-card-subtitle">\(escape(card.subtitle))</p>
-          <footer class="transcript-actions">
-            <button type="button" data-testid="tool-card-copy" data-copy-id="\(escape(copyID))">\(escape(copyActionLabel(for: card)))</button>
-          </footer>
-          \(renderToolArtifacts(card.artifacts))
-          \(renderToolTextPreviews(card.artifacts))
-          \(renderToolDocumentPreviews(card.artifacts))
-          \(renderToolImagePreviews(card.artifacts))
-          \(renderToolDetails(card))
-        </article>
-        """
-    }
-
-    private static func renderExecutionContextChip(
-        _ context: ExecutionContextSurface?,
-        testID: String
-    ) -> String {
-        guard let context else { return "" }
-        let title: String
-        switch context.kind {
-        case .local:
-            title = context.label
-        case .sshRemote:
-            title = "\(context.label) · \(context.detail)"
-        }
-        return """
-        <span class="execution-context-chip" data-testid="\(escape(testID))" data-execution-context-kind="\(escape(context.kind.rawValue))">\(escape(title))</span>
-        """
-    }
-
-    private static func copyActionLabel(for card: ToolCardState) -> String {
-        if card.outputJSON != nil {
-            return "Copy output"
-        }
-        if card.inputJSON != nil {
-            return "Copy input"
-        }
-        return "Copy"
-    }
-
-    private static func renderToolDetails(_ card: ToolCardState) -> String {
-        guard card.inputJSON != nil || card.outputJSON != nil else { return "" }
-        let isOpen = card.opensDetailsByDefault
-        return """
-        <details data-testid="tool-card-details"\(isOpen ? " open" : "")>
-          <summary>\(isOpen ? "Hide details" : "Show raw details")</summary>
-          \(card.inputJSON.map { #"<pre data-testid="tool-card-input">\#(escape($0))</pre>"# } ?? "")
-          \(card.outputJSON.map { #"<pre data-testid="tool-card-output">\#(escape($0))</pre>"# } ?? "")
-        </details>
-        """
-    }
-
-    private static func renderToolArtifacts(_ artifacts: [ToolArtifactState]) -> String {
-        guard !artifacts.isEmpty else { return "" }
-        let chips = artifacts.map { artifact in
-            let href = artifact.href.map { #" href="\#(escape($0))""# } ?? ""
-            return """
-            <a class="artifact-chip" data-testid="tool-card-artifact" data-kind="\(escape(artifact.kind.rawValue))"\(href)>
-              <strong data-testid="tool-card-artifact-label">\(escape(artifact.label))</strong>
-              <small data-testid="tool-card-artifact-detail">\(escape(artifact.detail))</small>
-            </a>
-            """
-        }.joined(separator: "\n")
-        return """
-        <div class="tool-artifacts" data-testid="tool-card-artifacts" aria-label="Artifacts">
-          \(chips)
-        </div>
-        """
-    }
-
-    private static func renderToolTextPreviews(_ artifacts: [ToolArtifactState]) -> String {
-        let textArtifacts = artifacts.filter(\.hasTextPreview)
-        guard !textArtifacts.isEmpty else { return "" }
-        let previews = textArtifacts.map { artifact in
-            """
-            <figure class="artifact-text-preview" data-testid="tool-card-text-preview">
-              <figcaption data-testid="tool-card-text-preview-label">\(escape(artifact.label))</figcaption>
-              <pre data-testid="tool-card-text-preview-content">\(escape(artifact.textPreview ?? ""))</pre>
-            </figure>
-            """
-        }.joined(separator: "\n")
-        return """
-        <div class="tool-artifact-text-previews" data-testid="tool-card-text-previews" aria-label="Text previews">
-          \(previews)
-        </div>
-        """
-    }
-
-    private static func renderToolDocumentPreviews(_ artifacts: [ToolArtifactState]) -> String {
-        let documentArtifacts = artifacts.filter(\.isDocumentPreview)
-        guard !documentArtifacts.isEmpty else { return "" }
-        let previews = documentArtifacts.compactMap { artifact -> String? in
-            guard let preview = artifact.documentPreview else { return nil }
-            let openLink = artifact.href.map {
-                #"<a data-testid="tool-card-document-preview-open" href="\#(escape($0))">Open</a>"#
-            } ?? ""
-            return """
-            <figure class="artifact-document-preview" data-testid="tool-card-document-preview" data-kind="\(escape(preview.kind.rawValue))">
-              <span class="artifact-document-icon" aria-hidden="true">\(documentIcon(for: preview.kind))</span>
-              <figcaption>
-                <small data-testid="tool-card-document-preview-type">\(escape(preview.typeLabel)) · \(escape(preview.extensionLabel))</small>
-                <strong data-testid="tool-card-document-preview-label">\(escape(artifact.label))</strong>
-                <small data-testid="tool-card-document-preview-detail">\(escape(preview.detail))</small>
-              </figcaption>
-              \(openLink)
-            </figure>
-            """
-        }.joined(separator: "\n")
-        guard !previews.isEmpty else { return "" }
-        return """
-        <div class="tool-artifact-document-previews" data-testid="tool-card-document-previews" aria-label="Document previews">
-          \(previews)
-        </div>
-        """
-    }
-
-    private static func renderToolImagePreviews(_ artifacts: [ToolArtifactState]) -> String {
-        let imageArtifacts = artifacts.filter(\.isImagePreview)
-        guard !imageArtifacts.isEmpty else { return "" }
-        let previews = imageArtifacts.compactMap { artifact -> String? in
-            guard let src = artifact.previewURL,
-                  let preview = artifact.imagePreview
-            else { return nil }
-            return """
-            <figure class="artifact-preview" data-testid="tool-card-image-preview" data-kind="image">
-              <img src="\(escape(src))" alt="\(escape(artifact.label))" loading="lazy">
-              <figcaption>
-                <small data-testid="tool-card-image-preview-type">\(escape(preview.typeLabel)) · \(escape(preview.extensionLabel))</small>
-                <strong data-testid="tool-card-image-preview-label">\(escape(artifact.label))</strong>
-                <small data-testid="tool-card-image-preview-detail">\(escape(preview.detail))</small>
-              </figcaption>
-            </figure>
-            """
-        }.joined(separator: "\n")
-        guard !previews.isEmpty else { return "" }
-        return """
-        <div class="tool-artifact-previews" data-testid="tool-card-image-previews" aria-label="Image previews">
-          \(previews)
-        </div>
-        """
-    }
-
-    private static func documentIcon(for kind: ToolArtifactDocumentKind) -> String {
-        switch kind {
-        case .appshot:
-            return "APP"
-        case .pdf:
-            return "PDF"
-        case .document:
-            return "DOC"
-        case .spreadsheet:
-            return "XLS"
-        case .presentation:
-            return "PPT"
-        }
-    }
-
     private static func renderTerminal(_ terminal: TerminalSurface) -> String {
         guard terminal.isVisible else { return "" }
         let entries = terminal.entries.isEmpty
@@ -592,7 +419,7 @@ public enum WorkspaceHTMLRenderer {
                   <header>
                     <span class="terminal-command-row">
                       <code>$ \(escape(entry.command))</code>
-                      \(renderExecutionContextChip(entry.executionContext, testID: "terminal-execution-context"))
+                      \(WorkspaceHTMLPrimitives.executionContextChip(entry.executionContext, testID: "terminal-execution-context"))
                     </span>
                     <span class="terminal-status \(terminalStatusClass(entry))" data-testid="terminal-status">\(escape(entry.statusLabel)) · \(escape(entry.exitCodeLabel))</span>
                   </header>
@@ -1029,11 +856,6 @@ public enum WorkspaceHTMLRenderer {
     }
 
     private static func escape(_ text: String) -> String {
-        text
-            .replacingOccurrences(of: "&", with: "&amp;")
-            .replacingOccurrences(of: "<", with: "&lt;")
-            .replacingOccurrences(of: ">", with: "&gt;")
-            .replacingOccurrences(of: "\"", with: "&quot;")
-            .replacingOccurrences(of: "'", with: "&#39;")
+        WorkspaceHTMLPrimitives.escape(text)
     }
 }
