+++ docs/CODE_QUALITY_AUDIT.md
@@ -26,6 +26,7 @@ The architecture is moving in the right direction: core state is value typed, pe
 | `Sources/QuillCodeApp/WorkspaceModel.swift` | A- | Command parsing, automation records/run drafts, terminal session construction, project registry transitions, review-comment planning, tool override composition, SSH Remote tool execution, browser location/state transitions, MCP surface state, MCP request parsing, MCP runtime/catalog/launch work, tool-card surface types, execution-context enrichment, thread seeding, thread lifecycle transitions, sidebar selection transitions, and sidebar bulk action planning now live in focused helpers; keep extracting pure surface/workflow builders before adding more parity commands. |
 | `Sources/QuillCodeApp/WorkspaceSwiftUIView.swift` | B+ | The shell is now mostly composition, state, and routing. Next step is moving remaining transcript/find/context-banner rendering or command-routing helpers out if they grow again. |
 | `Sources/QuillCodeApp/WorkspaceSurface.swift` | A- | Surface assembly is still large, but settings copy/compatibility, runtime issue classification, model catalog presentation, command construction, command palette ranking, review diff construction, context banner estimation, and transcript message projection are now extracted into focused files. Next step is extracting additional surface-family builders only when behavior grows. |
+| `Sources/QuillCodeApp/WorkspaceHTMLRenderer.swift` | A- | Static HTML harness rendering is still broad, but tool-card/artifact preview HTML now delegates to `WorkspaceHTMLToolCardRenderer`, and shared escaping/context chips live in `WorkspaceHTMLPrimitives`. Next step is extracting another pane family only when renderer drift appears. |
 | `Sources/quill-code-desktop/QuillCodeDesktopApp.swift` | A- | App scene composition is now small and declarative. Keep it limited to window/menu-bar wiring and root-view routing. |
 | `Sources/quill-code-desktop/QuillCodeDesktopController.swift` | A- | Desktop controller is now mostly UI/workspace routing. Pasteboard feedback and project-import resolution now live in focused coordinators; keep future desktop protocol/workflow details out of the controller. |
 | `Sources/QuillCodeAgent/Agent.swift` | A- | Good test coverage; keep tool continuation limits and transcript filtering explicit. |
@@ -782,3 +783,20 @@ Code quality changes:
 Remaining risk:
 
 - `WorkspaceSurface.swift` still carries many small value records. That remains acceptable while those records are plain Codable data, but any record with compatibility decoding or presentation helpers should move into a focused surface-family file before new Codex-parity behaviors land.
+
+## 2026-06-23 HTML Tool Card Renderer Pass
+
+Overall grade after this slice: **A- foundation, A static HTML tool-card boundary**.
+
+Static HTML tool-card rendering, artifact chips, text previews, document previews, image previews, raw details, copy labels, and document icon labels moved out of `WorkspaceHTMLRenderer.swift` into `WorkspaceHTMLToolCardRenderer.swift`. Shared HTML escaping and execution-context chip rendering moved into `WorkspaceHTMLPrimitives.swift`, so tool cards and terminal rows no longer maintain separate context-chip markup.
+
+Code quality changes:
+
+- Added `WorkspaceHTMLToolCardRenderer` as the focused owner for tool-card HTML.
+- Added `WorkspaceHTMLPrimitives` for shared escaping and execution-context chip markup.
+- Reduced the static HTML renderer by roughly 190 lines while keeping the public `WorkspaceHTMLRenderer.render(_:)` contract unchanged.
+- Added a parity gate so artifact, preview, details, document-icon, escaping, and execution-context chip markup do not drift back into the monolithic renderer.
+
+Remaining risk:
+
+- `WorkspaceHTMLRenderer.swift` still owns several pane renderers because it is the static harness composition point. Keep extracting whole pane families only when they gain enough behavior to risk drifting from SwiftUI.
