+++ docs/DECISIONS.md
@@ -155,6 +155,7 @@
 - Desktop transcript copy behavior lives in `QuillCodeDesktopCopyCoordinator`. The controller owns only the visible copied-item state, while the coordinator owns blank-copy rejection, pasteboard mutation through `QuillCodePasteboardWriting`, and the transient feedback duration. This keeps AppKit pasteboard details out of workspace routing.
 - Desktop project import result handling lives in `QuillCodeDesktopProjectImportCoordinator`. The controller owns only SwiftUI importer presentation state, while the coordinator owns `fileImporter` result parsing, URL normalization, and real directory validation before the controller adds a project.
 - Tool-card and artifact presentation models live in `QuillCodeToolCardSurface.swift`, not in `WorkspaceModel.swift`. `WorkspaceModel` may assemble tool cards from events, but status/density, artifact kind detection, document/image/text-preview metadata, and default expansion behavior are surface concerns with their own parity gate.
+- Static HTML tool-card rendering lives in `WorkspaceHTMLToolCardRenderer.swift`. `WorkspaceHTMLRenderer.swift` remains the composition point for the Playwright/static harness, while artifact chips, text previews, image/document previews, raw detail blocks, and copy labels stay beside the tool-card HTML renderer. Shared HTML escaping and execution-context chip markup live in `WorkspaceHTMLPrimitives.swift` so terminal rows and tool cards cannot drift.
 - Browser comment, snapshot, and browser panel state live in `QuillCodeBrowserSurface.swift`; browser address resolution lives in `WorkspaceBrowserLocationResolver.swift`; and browser history/page/comment state transitions live in `WorkspaceBrowserEngine`. The workspace model owns async fetch orchestration, `lastError`, and top-bar refreshes, while reusable browser state contracts, URL-resolution rules, and pure state transitions stay focused and guarded by parity tests.
 - Review comment payload state and event planning live in `WorkspaceReviewCommentPlanner.swift`, not in `WorkspaceModel.swift`. The planner owns path/text trimming, visible review-file validation, line-range normalization, line-kind checks, summary copy, and payload encoding. The workspace model owns only selected-thread validation, event append, persistence, and top-bar refresh.
 - Agent tool execution override composition lives in `WorkspaceToolExecutionOverrideCombiner.swift`, not in `WorkspaceModel.swift`. The combiner owns the precedence order Plan, Remote Project, Browser, Computer Use, Memory, MCP, while the workspace model only creates each optional executor and assigns the combined override to the runner.
